+++ packages/web3-providers/src/Twitter/apis/getUserViaTwitterIdentity.ts
@@ -1,7 +1,7 @@
 import urlcat from 'urlcat'
 import { fetchCachedJSON } from '../../helpers/fetchJSON.js'
-import { staleCached } from '../../helpers/fetchCached.js'
-import { Duration, Expiration } from '../../entry-helpers.js'
+import { Duration, staleCached } from '../../helpers/fetchCached.js'
+import { Expiration } from '../../helpers/fetchSquashed.js'
 import type { TwitterBaseAPI } from '../../entry-types.js'
 
 const TWITTER_IDENTITY_URL = 'https://mr8asf7i4h.execute-api.us-east-1.amazonaws.com/prod/twitter-identity'
